Carico esercizio completo

+++ database/migrations/2021_05_19_142340_create_holidays_table.php
@@ -19,7 +19,7 @@ class CreateHolidaysTable extends Migration
             $table->tinyInteger('days');
             $table->string('name_hotel');
             $table->string('name_operator');
-            $table->smallInteger('price');
+            $table->float('price');
             $table->tinyInteger('people');
             $table->timestamps();
         });
